missed a file
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18527 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/HUDuiTypeIn.h
@@ -61,7 +61,7 @@ class HUDuiTypeIn : public HUDuiControl {
 			~HUDuiTypeIn();
 
     void		setObfuscation(bool on);
-    int			getMaxLength() const;
+    size_t		getMaxLength() const;
     std::string		getString() const;
 
     void		setMaxLength(size_t);
